pyproject.toml
[project]
name = "scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.28.0",
    "email-validator>=2.0",
    "fastapi>=0.109.0",
    "numpy>=1.24.0",
    "bcrypt>=4.0",
    "pillow>=10.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.6",
    "redis>=5.0",
    "sqlalchemy[asyncio]>=2.0",
    "uvicorn[standard]>=0.27.0",
    "ruff>=0.14.14",
    "ty>=0.0.14",
    "structlog>=25.5.0",
    "prometheus-client>=0.24.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation-redis>=0.60b1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-logging>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "pyjwt>=2.11.0",
    "slowapi>=0.1.9",
    "alembic>=1.18.4",
    "celery>=5.6.2",
]

[tool.ruff]
line-length = 88
fix = true
target-version = "py313"

[tool.ruff.lint]
select = [
    "N", # pyupgrade
    "C4", # flake8-bugbear
    "RUF", # Ruff-specific rules
    "E", # ошибки стиля (pycodestyle)
    "F", # логические ошибки (pyflakes)
    "B", # баг-паттерны (bugbear)
    "I", # импорты
    "UP", # обновление синтаксиса
    "SIM", # упрощения
    "PERF", # производительность
]

ignore = [
    "B008", # Depends/Query/File в FastAPI — намеренное использование
]

[tool.ty.environment]
# Use project root so "src" is the top-level package and ..core etc. resolve.
root = ["."]

[tool.ty.analysis]
allowed-unresolved-imports = []

# Relax strict checks for Starlette/FastAPI app wiring and generic services
[[tool.ty.overrides]]
include = ["src/main.py", "src/core/middleware.py"]
[tool.ty.overrides.rules]
invalid-argument-type = "ignore"

[[tool.ty.overrides]]
include = ["src/repositories/**", "src/services/base.py", "src/services/advice_service.py", "src/services/cause_service.py"]
[tool.ty.overrides.rules]
# Generic T/ReadSchema bounds (DeclarativeBase, SchemaBase) have these attributes at runtime
possibly-missing-attribute = "ignore"
unresolved-attribute = "warn"
invalid-return-type = "error"

[tool.lint.isort]
known-first-party = ["scratch", "src"]

